Renaming and openfoam wrapper and removing alternate caps in interfoam and pimplefoam wrappers.
