--- coupling_components/solver_wrappers/openfoam/CoCoNuT_pimpleFoam/Make/files
@@ -1,3 +1,0 @@
-CoCoNuT_pimpleFoam.C
-
-EXE = $(FOAM_USER_APPBIN)/CoCoNuT_pimpleFoam
