+++ coupling_components/solver_wrappers/openfoam/coconut_interFoam/Make/linux64IccDPInt32Opt/variables
