+++ coupling_components/solver_wrappers/openfoam/coconut_pimpleFoam/Make/files
@@ -1,0 +1,3 @@
+coconut_pimpleFoam.C
+
+EXE = $(FOAM_USER_APPBIN)/coconut_pimpleFoam
